requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.4
black==19.10b0
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
distlib==0.3.1
Django==2.2
django-autocomplete-light==3.5.1
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-environ==0.4.5
django-js-asset==1.2.2
filelock==3.0.12
flake8==3.8.3
gunicorn==20.0.4
identify==1.4.25
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.782
mypy-extensions==0.4.3
nodeenv==1.4.0
pathspec==0.8.0
pep8==1.7.1
Pillow==7.2.0
pre-commit==2.6.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.10
virtualenv==20.0.30
whitenoise==5.2.0
wrapt==1.12.1
